+++ site/src/docs/docs.11tydata.js
@@ -1,0 +1,4 @@
+export default {
+  layout: "docs.njk",
+  tags: ["docs"]
+};